requirements.txt
pystan==2.19.1.1
numpy==1.21.4
prophet
appdirs==1.4.4
appnope==0.1.2
backcall==0.2.0
beautifulsoup4==4.9.3
bidict==0.21.4
bs4==0.0.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
chart-studio==1.1.0
click==8.0.3
cmdstanpy==0.9.68
convertdate==2.3.2
cssselect==1.1.0
cycler==0.11.0
Cython==0.29.24
decorator==5.1.0
ephem==4.1
fake-useragent==0.1.11
Flask==1.1.2
Flask-SocketIO==4.3.2
googlesearch-python==1.0.1
hijri-converter==2.2.2
holidays==0.11.3.1
idna==2.10
imageio==2.10.2
importlib-metadata==4.8.1
ipython==7.29.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jsonpickle==2.0.0
kiwisolver==1.3.2
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
lxml==4.6.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
networkx==2.6.3
nltk==3.6.5
opencv-python==4.5.4.58
pandas==1.3.4
parse==1.19.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.4.0
prompt-toolkit==3.0.22
ptyprocess==0.7.0
pyee==8.2.2
Pygments==2.10.0
PyMeeus==0.5.11
pyparsing==3.0.4
pyppeteer==0.2.6
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQtWebEngine==5.15.5
PyQtWebEngine-Qt5==5.15.2
pyquery==1.4.3
pytesseract==0.3.8
python-dateutil==2.8.2
python-engineio==3.14.2
python-socketio==4.6.1
pytz==2021.3
pyvis==0.1.9
PyWavelets==1.1.1
regex==2021.11.2
requests==2.25.1
requests-html==0.10.0
retrying==1.3.3
scikit-image==0.18.3
scipy==1.7.1
setuptools-git==1.2
six==1.16.0
soupsieve==2.3
tenacity==8.0.1
tifffile==2021.11.2
tqdm==4.62.3
traitlets==5.1.1
ujson==4.2.0
urllib3==1.26.7
w3lib==1.22.0
wcwidth==0.2.5
websockets==9.1
Werkzeug==2.0.2
zipp==3.6.0
flask-dropzone==1.6.0
flask-login==0.6.0
wtforms==3.0.0
flask-wtf==1.0.1
flask-bcrypt==1.0.1
flask-sqlalchemy==2.5.1
email-validator==1.1.3
